requirements.txt
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0
datasets>=2.14.0
accelerate>=0.20.0
scipy>=1.10.0
numpy>=1.24.0
puzpy>=0.3.0
pyyaml>=6.0
tqdm>=4.65.0
wordsegment>=1.3.1
faiss-cpu>=1.7.4
wandb>=0.15.0
